mostly implemented separate view for saving edits, and fixed a bug in view edit
