+++ packages/views/order_edit/order_edit.py
@@ -10,12 +10,13 @@ def order_edit_view(request, partner_name, order_id):
     order = Order.objects.get(id=order_id)
 
     cookies = request.COOKIES
+    make_cookies_make_sense(cookies)
 
-    if request.POST.get('rate'):
-        save_changes_to_cookies(request, cookies, order)
+    if request.POST:
+        cookies = changed_cookies(request, cookies, order)
     
-    new_packages_cookie = request.COOKIES.get(str(order_id)+'_new_packages')
-    updated_packages_cookie = request.COOKIES.get(str(order_id)+'_updated_packages')
+    new_packages_cookie = cookies.get(str(order_id)+'_new_packages')
+    updated_packages_cookie = cookies.get(str(order_id)+'_updated_packages')
 
     packages = []
 
@@ -27,8 +28,6 @@ def order_edit_view(request, partner_name, order_id):
     for package in list(order.related_packages()):
         if package not in packages:
             packages.append(package)
-
-    packages.reverse()
 
     context = {
         'packages': packages,
@@ -42,66 +41,27 @@ def order_edit_view(request, partner_name, order_id):
 
     response = render(request, 'packages/order_edit.html', context)
 
-    if updated_packages_cookie:
-        response.set_cookie(str(order.id)+'_updated_packages', updated_packages_cookie)
-    if new_packages_cookie:
-        response.set_cookie(str(order.id)+'_new_packages', new_packages_cookie)
-
-    if request.POST.get('save'):
-        if updated_packages_cookie:
-            updated_packages = update_saved_packages(updated_packages_cookie)
-            send_update_email(updated_packages)
-        if new_packages_cookie:
-            new_packages = create_saved_packages(new_packages_cookie)
-            send_new_email(new_packages)
-        response.delete_cookie(str(order_id)+'_updated_packages')
-        response.delete_cookie(str(order_id)+'_new_packages')
+    response.set_cookie(str(order.id)+'_updated_packages', updated_packages_cookie)
+    response.set_cookie(str(order.id)+'_new_packages', new_packages_cookie)
 
     return response
+
+
+def make_cookies_make_sense(cookies):
+    #this caused a lot of confusion
+    for key, value in cookies.items():
+        if value == 'None':
+            cookies[key] = None
 
 
 def json_to_packages(json):
     return [des.object for des in serializers.deserialize('json', json)]
 
-def update_saved_packages(updated_packages_cookie):
-    packages = serializers.deserialize('json', updated_packages_cookie)
 
-    for package in packages:
-        actual_package = Package.objects.filter(id=package.id)
-        actual_package.update(origin=package.origin, destination=package.destination, rate=package.rate,
-                              full_name=package.full_name, phone_number=package.phone_number, notes=package.notes)
-
-    return packages
-
-
-def create_saved_packages(new_packages_cookie):
-    packages = serializers.deserialize('json', new_packages_cookie)
-
-    for package in packages:
-        package = package.object
-        Package.objects.create(order=package.order, origin=package.origin, destination=package.destination,
-                               rate=package.rate, full_name=package.full_name, phone_number=package.phone_number,
-                               notes=package.notes)
-
-    return packages
-
-
-def send_update_email(packages):
-    #TODO:
-    #
-    pass
-
-
-def send_new_email(packages):
-    #TODO:
-    #
-    pass
-
-
-def save_changes_to_cookies(request, cookies, order):
+def changed_cookies(request, cookies, order):
     new_packages_cookie = cookies.get(str(order.id)+'_new_packages')
     updated_packages_cookie = cookies.get(str(order.id)+'_updated_packages')
-
+    
     origin_address = get_or_create_origin_address(request)
     destination_address = get_or_create_destination_address(request)
 
@@ -116,18 +76,16 @@ def save_changes_to_cookies(request, cookies, order):
                       full_name=full_name, order=order, notes=notes)
     if package_id:
         package.id = package_id
-        json = updated_packages_cookie
-        json = add_package_to_json(package, json)
+        json = add_package_to_json(package, updated_packages_cookie)
 
         cookies[str(order.id)+'_updated_packages'] = json
 
     else:
-        json = new_packages_cookie
-        print('\n\nbefore: ' + str(new_packages_cookie) + '\n' + '\n')
-        json = add_package_to_json(package, json)
-        print('after: ' + str(json) + '\n' + '\n')
+        json = add_package_to_json(package, new_packages_cookie)
 
         cookies[str(order.id)+'_new_packages'] = json
+
+    return cookies
 
 
 def get_or_create_destination_address(request):
@@ -165,8 +123,6 @@ def add_package_to_json(package, json):
     if json:
         packages = json_to_packages(json)
         packages.append(package)
-        print(type(package))
-        print(type(packages[0]))
     else:
         packages = [package]
